Process serialization asynchronously.
